requirements.txt
# Core dependencies
azure-ai-formrecognizer>=3.2.0
chromadb>=0.4.0
docker>=6.1.0
docker-compose>=1.29.0
python-dotenv>=1.0.0
requests>=2.31.0

# Vector storage and embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0

# Document processing
PyMuPDF>=1.22.0
Pillow>=9.5.0
python-poppler>=0.3.0
pdf2image>=1.16.3

# Graph RAG
graphrag>=0.1.0
networkx>=3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pylint>=2.17.0
